+++ irs-api/src/main/java/org/eclipse/tractusx/irs/edc/EdcCallbackController.java
@@ -27,7 +27,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.eclipse.dataspaceconnector.spi.types.domain.edr.EndpointDataReference;
 import org.eclipse.tractusx.irs.IrsApplication;
-import org.eclipse.tractusx.irs.util.JsonUtil;
+import org.eclipse.tractusx.irs.util.Masker;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -44,13 +44,13 @@ import org.springframework.web.bind.annotation.RestController;
 public class EdcCallbackController {
 
     private final EndpointDataReferenceStorage storage;
-    private final JsonUtil jsonUtil;
 
     @PostMapping("/endpoint-data-reference")
     public void receiveEdcCallback(final @RequestBody EndpointDataReference dataReference) {
-        log.debug("Received EndpointDataReference: {}", jsonUtil.asString(dataReference));
+        log.debug("Received EndpointDataReference with ID {} and endpoint {}", dataReference.getId(),
+                dataReference.getEndpoint());
         final var contractAgreementId = dataReference.getProperties().get("cid");
         storage.put(contractAgreementId, dataReference);
-        log.info("Endpoint Data Reference received and cached for agreement: {}", contractAgreementId);
+        log.info("Endpoint Data Reference received and cached for agreement: {}", Masker.mask(contractAgreementId));
     }
 }