+++ charts/irs-helm/CHANGELOG.md
@@ -5,6 +5,25 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add minio resource limits 
+- Extended configmap and values.yaml with catalog cache configuration  
+  You can configure the EDC catalog caching configuration like this:
+  ```
+  edc:
+    catalog:
+      cache:
+        enabled: true
+        ttl: P1D
+        maxCachedItems: 64000
+  ```
+
+## [5.0.6] - 2023-03-28
+### Added
+- Added config parameter for SemanticsHub request page size when retrieving all models. Can be used to fine tune requests. Default: 100 items per page
+
+### Changed
+- Updated default path in template for `edc.controlplane.endpoint.data` to match EDC 0.3.0 management endpoint `/api/v1/management`
 
 ## [5.0.5] - 2023-03-20
 ### Changed